backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
reportlab==4.0.4
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional dependencies for enhanced functionality
# pdfplumber==0.9.0  # Alternative PDF parser
# spacy==3.6.1       # For advanced NLP
# nltk==3.8.1        # For text processing
# pandas==2.1.0      # For data analysis
